pyproject.toml
[tool.poetry]
name = "take-home-test"
version = "0.1.0"
description = ""
authors = ["Marco Rodríguez"]

[tool.poetry.dependencies]
python = "^3.9"
typer = "^0.6.1"
sqlmodel = "^0.0.6"
psycopg2-binary = "^2.9.3"
click = "^8.1.3"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
typer = "^0.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
